+++ web-admin/src/routes/[organization]/[project]/+layout.svelte
@@ -1,37 +1,3 @@
-<script context="module" lang="ts">
-  const PollTimeWhenProjectDeploymentPending = 1000;
-  const PollTimeWhenProjectDeploymentError = 5000;
-  const PollTimeWhenProjectDeploymentOk = RUNTIME_ACCESS_TOKEN_DEFAULT_TTL / 2; // Proactively refetch the JWT before it expires
-
-  const baseGetProjectQueryOptions: QueryObserverOptions<
-    V1GetProjectResponse,
-    RpcStatus
-  > = {
-    cacheTime: Math.min(RUNTIME_ACCESS_TOKEN_DEFAULT_TTL, 1000 * 60 * 5), // Make sure we don't keep a stale JWT in the cache
-    refetchInterval: (data) => {
-      switch (data?.prodDeployment?.status) {
-        case V1DeploymentStatus.DEPLOYMENT_STATUS_PENDING:
-          return PollTimeWhenProjectDeploymentPending;
-        case V1DeploymentStatus.DEPLOYMENT_STATUS_ERROR:
-          return PollTimeWhenProjectDeploymentError;
-        case V1DeploymentStatus.DEPLOYMENT_STATUS_OK:
-          return PollTimeWhenProjectDeploymentOk;
-      }
-    },
-    refetchOnMount: true,
-    refetchOnReconnect: true,
-    refetchOnWindowFocus: true,
-    select: (data: V1GetProjectResponse) => {
-      if (data?.prodDeployment?.runtimeHost) {
-        data.prodDeployment.runtimeHost = fixLocalhostRuntimePort(
-          data.prodDeployment.runtimeHost,
-        );
-      }
-      return data;
-    },
-  };
-</script>
-
 <script lang="ts">
   import { page } from "$app/stores";
   import {
@@ -39,23 +5,20 @@
     createAdminServiceGetCurrentUser,
     createAdminServiceGetDeploymentCredentials,
     createAdminServiceGetProject,
-    type RpcStatus,
-    type V1GetProjectResponse,
   } from "@rilldata/web-admin/client";
   import { isProjectPage } from "@rilldata/web-admin/features/navigation/nav-utils";
   import ProjectBuilding from "@rilldata/web-admin/features/projects/ProjectBuilding.svelte";
   import ProjectTabs from "@rilldata/web-admin/features/projects/ProjectTabs.svelte";
+  import { baseGetProjectQueryOptions } from "@rilldata/web-admin/features/projects/status/selectors";
   import RedeployProjectCta from "@rilldata/web-admin/features/projects/RedeployProjectCTA.svelte";
   import { createAdminServiceGetProjectWithBearerToken } from "@rilldata/web-admin/features/public-urls/get-project-with-bearer-token";
   import { viewAsUserStore } from "@rilldata/web-admin/features/view-as-user/viewAsUserStore";
   import ErrorPage from "@rilldata/web-common/components/ErrorPage.svelte";
   import { metricsService } from "@rilldata/web-common/metrics/initMetrics";
   import RuntimeProvider from "@rilldata/web-common/runtime-client/RuntimeProvider.svelte";
-  import { RUNTIME_ACCESS_TOKEN_DEFAULT_TTL } from "@rilldata/web-common/runtime-client/constants";
   import type { HTTPError } from "@rilldata/web-common/runtime-client/fetchWrapper";
   import { fixLocalhostRuntimePort } from "@rilldata/web-common/runtime-client/fix-localhost-runtime-port";
   import type { AuthContext } from "@rilldata/web-common/runtime-client/runtime-store";
-  import type { QueryObserverOptions } from "@tanstack/svelte-query";
 
   const user = createAdminServiceGetCurrentUser();
